requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
flask-limiter
rich
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
Flask-Dance
blinker
pytest
psycopg2-binary
gevent